requirements.txt
Flask
websocket-client
pandas
requests
plotly

